refactor(lib, config): split config to file feat(config): now config error will auto convert to ScriptError
